Added gap function for pan
