--- creditcards/fd-card/fd.js
@@ -1,8 +1,0 @@
-import groupCharacters from '../../common/common-fn.js';
-
-const addGaps = () => {
-  const inputField = document.querySelector('.char-gap-4 input');
-  groupCharacters(inputField, 4);
-};
-
-export default addGaps;
